Split generic type-reference validation
